+++ src/app/static/src/app/components/plots/bubble/SizeLegend.vue
@@ -19,8 +19,7 @@
                 </div>
             </div>
             <map-adjust-scale class="legend-element legend-adjust map-control" name="size" :step="scaleStep"
-                              :show="showAdjust" :scale="sizeScale" @update="update" :metadata="metadata"
-                              :hide-static-custom="true" :hide-static-default="true">
+                              :show="showAdjust" :scale="sizeScale" @update="update" :metadata="metadata">
             </map-adjust-scale>
         </div>
     </l-control>
@@ -103,7 +102,9 @@
                 return this.width / 2;
             },
             circles: function () {
-                if (this.indicatorRange.min == this.indicatorRange.max) {
+                if (isNaN(this.indicatorRange.min) || isNaN(this.indicatorRange.max)) {
+                    return [];
+                } else if (this.indicatorRange.min == this.indicatorRange.max) {
                     // only one value in range - show max circle only
                     return [this.circleFromRadius(this.maxRadius, this.indicatorRange.max, false)];
                 } else {