+++ src/app/static/src/tests/components/plots/mapAdjustScale.test.ts
@@ -153,40 +153,13 @@ describe("MapAdjustScale component", () => {
         expect((wrapper.find("#custom-max-input").element as HTMLInputElement).disabled).toBe(true);
     });
 
-    it("renders as expected when hide props both false", () => {
+    it("renders static scale controls as expected", () => {
         const wrapper = mount(MapAdjustScale, {store, propsData});
 
         expect(wrapper.find(".static-container").exists()).toBe(true);
         expect(wrapper.find(".static-default").exists()).toBe(true);
         expect(wrapper.find(".static-custom").exists()).toBe(true);
         expect(wrapper.find(".static-custom-values").exists()).toBe(true);
-    });
-
-    it("renders as expected when hide static defaults", () => {
-        const wrapper = mount(MapAdjustScale, {store, propsData: {...propsData, hideStaticDefault: true}});
-
-        expect(wrapper.find(".static-container").exists()).toBe(true);
-        expect(wrapper.find(".static-default").exists()).toBe(false);
-        expect(wrapper.find(".static-custom").exists()).toBe(true);
-        expect(wrapper.find(".static-custom-values").exists()).toBe(true);
-    });
-
-    it("renders as expected when hide static custom", () => {
-        const wrapper = mount(MapAdjustScale, {store, propsData: {...propsData, hideStaticCustom: true}});
-
-        expect(wrapper.find(".static-container").exists()).toBe(true);
-        expect(wrapper.find(".static-default").exists()).toBe(true);
-        expect(wrapper.find(".static-custom").exists()).toBe(false);
-        expect(wrapper.find(".static-custom-values").exists()).toBe(false);
-    });
-
-    it("renders as expected when hide static default and custom", () => {
-        const wrapper = mount(MapAdjustScale, {store, propsData: {...propsData, hideStaticDefault: true, hideStaticCustom: true}});
-
-        expect(wrapper.find(".static-container").exists()).toBe(false);
-        expect(wrapper.find(".static-default").exists()).toBe(false);
-        expect(wrapper.find(".static-custom").exists()).toBe(false);
-        expect(wrapper.find(".static-custom-values").exists()).toBe(false);
     });
 
     it("emits update event when type changes", () => {