+++ NEWS.md
@@ -1,3 +1,19 @@
+# hint 1.4.0
+
+* Include metadata default scale for bubble size
+
+# hint 1.3.0
+
+* Don't show ADR key to user after it has been entered.
+
+# hint 1.2.1
+
+* Fetch all relevant datasets from ADR without paging
+
+# hint 1.2.0
+
+* Support custom value ranges for bubble plot scales.
+
 # hint 1.1.3
 
 * Display text on map when no data in selections.